v26
Move some thread related code to job.
